# Conflicts:
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/app_views/CreateMealPage.kt
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/app_views/RecipeEditorPage.kt
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/app_views/SearchPage.kt
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/widgets/DishItemButton.kt
#	app/src/main/java/com/frontend/nutricheck/client/ui/view/widgets/MealBlock.kt
#	app/src/main/java/com/frontend/nutricheck/client/ui/view_model/recipe/RecipeOverviewViewModel.kt

+++ app/src/main/java/com/frontend/nutricheck/client/model/data_sources/data/MealItem.kt
@@ -3,5 +3,5 @@ package com.frontend.nutricheck.client.model.data_sources.data
 interface MealItem {
     val mealId: String
     val quantity: Double
-    val servings: Int
+    val servings: Double
 }
